Whatsapp integrations without memory, includes wpp loger/sender/parser, route GET POST, interface, some name changes for better understanding
