+++ app/api/chat/route.ts
@@ -1,14 +1,17 @@
 // app/api/chat/route.ts
 
 import { NextRequest, NextResponse } from "next/server";
-import { handleChat } from "@/lib/chatbot-handlers/central-chatbot-processor";   // <— your helper
+import { processIncomingMessage } from "@/lib/chatbot-handlers/central-conversation-flow-decider";   // <— your helper
 
 // POST  →  chat endpoint used by the UI
 export async function POST(req: NextRequest) {
+  console.log("[API Route - /api/chat] Received POST request from webhook/internal call"); 
   try {
     const { history = [] } = await req.json();        // expects { history:[...] }
-    const newHistory = await handleChat(history);     // call bot logic
-    return NextResponse.json({ history: newHistory });
+    console.log("[API Route - /api/chat] Calling processIncomingMessage with history:", JSON.stringify(history, null, 2)); 
+    const result = await processIncomingMessage(history);     // call bot logic
+    console.log("[API Route - /api/chat] processIncomingMessage result:", JSON.stringify(result, null, 2)); 
+    return NextResponse.json({ history: result.messages }); // Use result.messages
   } catch (err) {
     console.error("chat route error:", err);
     return NextResponse.json(
