+++ app/api/webhook/route.ts
@@ -12,7 +12,13 @@
  */
 
 import { NextRequest, NextResponse } from "next/server";
-import { logIncomingMessage, WebhookAPIBody } from "@/lib/llm-actions/whatsapp/whatsapp-message-logger";
+import { logIncomingMessage, WebhookAPIBody } from "@/lib/chatbot-handlers/whatsapp/whatsapp-message-logger";
+import { parseWhatsappMessage } from "@/lib/chatbot-handlers/whatsapp/whatsapp-payload-parser";
+import { processIncomingMessage } from "@/lib/chatbot-handlers/central-conversation-flow-decider";
+import { WhatsappSender } from "@/lib/chatbot-handlers/whatsapp/whatsapp-message-sender";
+import { BotResponse } from "@/lib/cross-channel-interfaces/standardized-conversation-interface";
+
+export const dynamic = "force-dynamic";
 
 // Read environment variables for security and feature toggling
 const WHATSAPP_VERIFY_TOKEN = process.env.WHATSAPP_VERIFY_TOKEN;
@@ -76,6 +82,63 @@ export async function POST(req: NextRequest) {
     const payload = (await req.json()) as WebhookAPIBody;
     console.log("Webhook POST request received. Processing payload...");
     logIncomingMessage(payload);
+
+    const parsedMessage = parseWhatsappMessage(payload);
+
+    if (parsedMessage && parsedMessage.text && parsedMessage.senderId) { 
+      console.log("[Webhook] Successfully parsed WhatsApp message:", JSON.stringify(parsedMessage, null, 2));
+
+      const historyForDecider = [
+        {
+          role: "user" as const, 
+          content: parsedMessage.text,
+        },
+      ];
+
+      console.log("[Webhook] Calling central decider with history:", JSON.stringify(historyForDecider, null, 2));
+      try {
+        const llmResult = await processIncomingMessage(historyForDecider);
+        console.log("[Webhook] Response from central decider:", JSON.stringify(llmResult.messages, null, 2));
+        
+        // Extract assistant's reply and send it back
+        if (llmResult.messages && llmResult.messages.length > 0) {
+          const lastMessage = llmResult.messages[llmResult.messages.length - 1];
+          if (lastMessage.role === 'assistant' && lastMessage.content) {
+            const assistantReplyText = lastMessage.content;
+            
+            const botResponse: BotResponse = {
+              text: assistantReplyText,
+            };
+
+            const sender = new WhatsappSender();
+            try {
+              console.log(`[Webhook] Attempting to send reply to ${parsedMessage.senderId}: "${assistantReplyText}"`);
+              await sender.sendMessage(parsedMessage.senderId, botResponse);
+              console.log("[Webhook] Reply successfully sent to WhatsApp user.");
+            } catch (sendError) {
+              console.error("[Webhook] Error sending reply via WhatsappSender:", sendError);
+            }
+          } else {
+            console.log("[Webhook] Last message from decider was not from assistant or had no content.");
+          }
+        } else {
+          console.log("[Webhook] No messages received from central decider.");
+        }
+
+      } catch (deciderError) {
+        console.error("[Webhook] Error calling central decider:", deciderError);
+      }
+
+    } else {
+      if (!parsedMessage) {
+        console.log("[Webhook] Failed to parse WhatsApp message or message was not actionable.");
+      } else if (!parsedMessage.text) {
+        console.log("[Webhook] Parsed WhatsApp message has no text content. Skipping decider call.");
+      } else if (!parsedMessage.senderId) {
+        console.log("[Webhook] Parsed WhatsApp message has no sender information (senderId). Skipping decider call.");
+      }
+    }
+
     return NextResponse.json({ status: "success" }, { status: 200 });
   } catch (error) {
     console.error("Error processing webhook POST request:", error);
